Make modules for HSM-sys
